api/requirements.txt
azure-functions
azurefunctions-extensions-http-fastapi
